refactor(server-core): colocate entity fakes under entity test dirs (#3009)
Phase 3 of issue #3009 — move domain test fakes out of
test/unit/core/helpers/ and inline spec definitions into their
entity test directory, mirroring the production hexagonal layout.

New colocated fakes (extracted from inline spec definitions):
- test/unit/core/entities/user/fake-repository.ts — FakeUserRepository
  was inline in 3 spec files (user/service, identity/registration,
  identity/password-recovery)
- test/unit/core/entities/client/fake-repository.ts — FakeClientRepository
- test/unit/core/entities/robot/fake-repository.ts — FakeRobotRepository
- test/unit/core/entities/permission/fake-repository.ts — FakePermissionRepository

Moved out of helpers/:
- helpers/fake-role-repository.ts   -> entities/role/fake-repository.ts
- helpers/fake-realm-repository.ts  -> entities/realm/fake-repository.ts
- helpers/fake-policy-repository.ts -> entities/policy/fake-repository.ts

Spec files updated to import from the new colocated paths; inline
class definitions removed. helpers/index.ts no longer re-exports
the moved entity fakes.

Kept as-is:
- OAuth2 verifier-spec FakeClientRepository (implements
  IOAuth2ClientRepository, distinct contract)
- helpers/fake-oauth2-*.ts (server-core-specific shape, not yet
  generic enough to live under entity test dirs)
- helpers/fake-mail-client.ts, helpers/fake-identity-*.ts,
  helpers/fake-session-manager.ts (workflow/orchestration helpers,
  not entity repositories)

Refs #3009
