+++ apps/server-core/test/unit/core/entities/policy/fake-repository.ts
@@ -6,7 +6,7 @@
  */
 
 import type { Policy } from '@authup/core-kit';
-import type { IPolicyRepository } from '../../../../src/core/entities/policy/types.ts';
+import type { IPolicyRepository } from '../../../../../src/core/entities/policy/types.ts';
 import { FakeEntityRepository } from '@authup/server-test-kit';
 
 export class FakePolicyRepository extends FakeEntityRepository<Policy> implements IPolicyRepository {
